+++ objc2/src/runtime.rs
@@ -170,8 +170,7 @@ impl Ivar {
 
     /// Returns the offset of self.
     pub fn offset(&self) -> isize {
-        let offset = unsafe { ffi::ivar_getOffset(self.as_ptr()) };
-        offset as isize
+        unsafe { ffi::ivar_getOffset(self.as_ptr()) }
     }
 
     /// Returns the `Encoding` of self.