+++ objc2/CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added `Id::retain_autoreleased` to allow following Cocoas memory management
   rules more efficiently.
 * Consistently allow trailing commas in `msg_send!`.
+* Added `msg_send_bool!`, a less error-prone version of `msg_send!` for
+  Objective-C methods that return `BOOL`.
 
 ### Changed
 * **BREAKING**: Changed signature of `Id::new` and `Id::retain` from